requirements.txt
web3
requests
Pillow